Fixed cylinders generation
The issue was in the indices which, being unsigned bytes, could hold up to a value of 255. Changed to GLuint.

+++ GraphicsProgramming/GraphicsProgramming/BasicShape.h
@@ -22,7 +22,7 @@ protected:
 	std::vector<float> vertices;
 	std::vector<float> normals;
 	std::vector<float> texCoordinates;
-	std::vector<GLubyte> indices;
+	std::vector<GLuint> indices;
 	//Origin of the shape (the origin is allways in the middle/center of the shape)
 	Vector3 origin;
 	//Scale of the shape
